[feat-#35] cors설정 및 분리
